update/ split CommentCard
